DNSDump/requirements.txt
dnspython>=2.4.0
rich>=13.0.0
textual>=0.50.0
